refactor(plugin-sdk): extract platform-agnostic HostCore from Host
Split the monolithic Host class into HostCore (platform-agnostic runtime)
and a thin Host shim that injects a QuickJS evaluator. Introduces the
PluginEvaluator interface as the single platform-specific seam, enabling
non-browser runtimes (mobile, Node tests) to implement their own evaluators
without duplicating ~1200 lines of hook/lifecycle logic.

- Add src/host-core.ts: HostCore class, PluginEvaluator interface, all
  hook factories (useState, useEffect, useRequests, useHeaders, openWindow,
  prove, done, doneWithOverlay), timeout management, and execution context
  registry
- Rewrite src/index.ts: Host extends HostCore; createQuickJSSandbox and
  createQuickJSEvaluator are the only QuickJS-specific code remaining;
  all backward-compat methods (createEvalCode, getPluginConfig,
  updateExecutionContext) preserved
- Add host-core Vite entry and ./host-core package.json export so mobile
  and other consumers can import HostCore without pulling in QuickJS
- Add test/host-core.test.ts: 24 Node.js unit tests using a new Function()
  mock evaluator (no browser/WASM required); covers lifecycle, hooks,
  openWindow, prove, done/doneWithOverlay, timeout, and addCapability

Closes #262
